Move utility methods to module.
