+++ lib/dynamic_scaffold/path_helper.rb
@@ -8,18 +8,18 @@ module DynamicScaffold
 
     module LocalInstanceMethods
       def dynamic_scaffold_path(action, options = {})
-      route = Rails.application.routes.routes.find do |r|
-        route_params = r.required_defaults
-        route_params[:controller] == params[:controller] && (route_params[:action] == action.to_s && r.name)
-      end
+        route = Rails.application.routes.routes.find do |r|
+          route_params = r.required_defaults
+          route_params[:controller] == params[:controller] && (route_params[:action] == action.to_s && r.name)
+        end
 
-      if route.nil?
-        raise DynamicScaffold::Error::RouteNotFound,
-          "Missing controller#action #{params[:controller]}##{action}"
-      end
+        if route.nil?
+          raise DynamicScaffold::Error::RouteNotFound,
+            "Missing controller#action #{params[:controller]}##{action}"
+        end
 
-      public_send("#{route.name}_path", options)
-    end
+        public_send("#{route.name}_path", options)
+      end
     end
   end
 end
